backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
beautifulsoup4
sentence-transformers
faiss-cpu
groq
pandas
python-dotenv
numpy
tabulate
